iceberg: split unresolved_partition_spec into its own lib
This is needed to be able to use this lib (that doesn't depend on most
of the iceberg proper) for config validation.
